+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.5] - 2025-07-22
+
 ### Added
 
 - Add support for loading plugin modules.
@@ -487,7 +489,8 @@ We registered a dummy package to PyPI to test the publishing. Therefore a versio
 Initial version
 
 
-[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.3.4...HEAD
+[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.3.5...HEAD
+[0.3.5]: https://github.com/s3r3t0/sereto/compare/v0.3.4...v0.3.5
 [0.3.4]: https://github.com/s3r3t0/sereto/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/s3r3t0/sereto/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/s3r3t0/sereto/compare/v0.3.1...v0.3.2